Now prioritize one endpoint's commits over the other.  This means that if two endpoints start transactions simultaneously, then one will always win.  (Before, both would back out and re-try, which could lead to live lock.)
Additionally, made it so that when make an inactive event active,
change its id so that out-of-order no accept messages will not
interfere.  (By out of order, I mean that because of the endpoint's
threading model, it could be that we send an event finished sentinel
before we send the no_accept message.  This would mean that the backed
out event would get rescheduled and mis-interpret the no-accept
message for it, rather than the postponed version.)