By permitting tuple returns, introduced a shortfall with returning externals.  Before, if you wanted to assign an external from a function, you'd just use:
extAssign some_func() to some_ext;

But if the function now returns more than one value, you can't get the
other values and still do the assignment of the external.  Added
syntax to handle this:

extAssign _ to some_ext, some_num = some_func();

Lexed, parsed, and type checked for it.  Still need to slice and emit
(as well as slice and emit for tuple return in general).