Starting to serialize and deserialize peered data to be transferred over the network.
I'm pretty sure that there will be a bug transferring peered lists of
lists/maps of lists/etc.  I may have to change the code so that peered
lists of lists/maps keep track of their references so that can know
when to replace reference with a new one and when to reuse reference.
See fixme at line 175 of waldoNetworkSerializer.py.  Also, clearly,
add more test cases.