Commit manager code used locks in such a way that if two events wanted to commit to the same state, then no other event would be able to commit to any other data.  This was because, to commit, an invalidation listener would have to take the global lock on the commit manager and then try to acquire, sequentially, each touched object's lock.
The purpose of this global lock was to ensure that we couldn't get
deadlock on a single host from trying to acquire commit locks on each
separate waldo reference.

This commit takes a different approach.  Instead of a global commit
lock, this commit orders objects by their uuids when trying to
commit.  Because all invalidation listeners acquire locks in the same
order, there cannot be deadlock.