Emitting map and list literals now.  There's a challenge though:
In both map and list literals, do not know whether to get_val on Waldo
variables or just pass in the variables (ie, unsure of whether to
assign by reference or by value).  Probably will update the language
so that can only have literal values in list and map literal
constructors.