Waldo needs to convert its objects back into regular python objects when external code calls into Waldo and expects a return value.
The previous version of the emitter had "de-waldoified" objects,
however, it did it outside of an event's transaction boundaries.  As
such, the runtime might have handed an incorrect value back to the
user.  This commit solves that problem.  Essentially, pass around a
sentinel indicating whether when we return from an internal method we
need to de-waldoify return values.  Then, while inside transaction,
dewaldoify.