Added a minor extension to a test case to ensure that when returning a function call, won't call the function multiple times.
Additionally, noticed a bug with actually emitting a call to a
sequence send message.  Essentially, require setting the message send
bit to be atomic with actually making the call.  As such, need to put
parens around ternary.  Otherwise, if return a call to a message send,
can get into trouble:

return msg_send();

Translates to:
  return internal_msg_send_call() if set_msg_send_bit else None
      if 0 in _ret_args_array
      else other stuff

Whereas we want:
  return (internal_msg_send_call() if set_msg_send_bit else None)
      if 0 in _ret_args_array
      else other stuff