can happen if commit is requested and then   a ---> b ---> c
a asks for commit.  b backs out and forwards commit request on to c.
c waits on active event map lock before receiving request for commit.
a tells b to back out and forwards the request to b to backout, which
forwards the request on to c.  Then, if c reads the backout before the
request to commit, we may get to this point.  Just ignore the request.